commit 7

+++ views/articles/liste.php
@@ -14,7 +14,10 @@
                 <p class="card-text">
                     <a class="text-align-right" href="index.php?route=details-article&id=<?= $article->id ?>">Voir plus</a>
                     <a class="text-align-right text-danger ml-2" href="index.php?route=modifier-article&id=<?= $article->id ?>">Modifier</a>
-                    <a class="text-align-right text-danger ml-2" href="index.php?route=supprimer-article&id=<?= $article->id ?>" onclick="return confirm('Etes-vous certains de vouloir supprimer ?')">Supprimer</a>
+
+                    <?php if (empty($_SESSION['identifiant']) && $_SESSION['identifiant'] == 'admin') : ?>
+                        <a class="text-align-right text-danger ml-2" href="index.php?route=supprimer-article&id=<?= $article->id ?>" onclick="return confirm('Etes-vous certains de vouloir supprimer ?')">Supprimer</a>
+                    <?php endif; ?>
                 </p>
             </div>
         </div>
